move glsl-js-mirrors to next to the glsl implementations
