+++ frontend/javascripts/oxalis/view/right-menu/mapping_info_view.js
@@ -27,6 +27,7 @@ import Cube from "oxalis/model/bucket_data_handling/data_cube";
 import Model from "oxalis/model";
 import message from "messages";
 import * as Utils from "libs/utils";
+import { jsConvertCellIdToHSLA } from "oxalis/shaders/segmentation.glsl";
 
 const { Option, OptGroup } = Select;
 
@@ -66,83 +67,9 @@ type State = {
   didRefreshMappingList: boolean,
 };
 
-// See the shader-side implementation of getElementOfPermutation in segmentation.glsl.js
-// for a detailed description.
-function getElementOfPermutation(
-  index: number,
-  sequenceLength: number,
-  primitiveRoot: number,
-): number {
-  const oneBasedIndex = (index % sequenceLength) + 1.0;
-  const isFirstElement = oneBasedIndex === 1.0;
-
-  // The GLSL implementation of pow is 2**(y * log2(x)) in which
-  // intermediate results can suffer from precision loss. The following
-  // code mimics this behavior to get a consistent coloring in GLSL and
-  // JS.
-  const imprecise = x => new Float32Array([x])[0];
-  function glslPow(x, y) {
-    return Math.floor(imprecise(2 ** (y * imprecise(Math.log2(x)))));
-  }
-  const sequenceValue = glslPow(primitiveRoot, oneBasedIndex) % sequenceLength;
-
-  // Only use sequenceLength if the requested element is the first of the sequence
-  // Otherwise, return the actual sequenceValue
-  return isFirstElement ? sequenceLength : sequenceValue;
-}
-
-// Input in [0,1]
-// Output in [0,1] for r, g and b
-function colormapJet(x: number): [number, number, number] {
-  const r = _.clamp(x < 0.89 ? (x - 0.35) / 0.31 : 1.0 - ((x - 0.89) / 0.11) * 0.5, 0, 1);
-  const g = _.clamp(x < 0.64 ? (x - 0.125) * 4.0 : 1.0 - (x - 0.64) / 0.27, 0, 1);
-  const b = _.clamp(x < 0.34 ? 0.5 + (x * 0.5) / 0.11 : 1.0 - (x - 0.34) / 0.31, 0, 1);
-
-  return [r, g, b];
-}
-
-// From: https://stackoverflow.com/a/54070620/896760
-// Input: r,g,b in [0,1], out: h in [0,360) and s,v in [0,1]
-function rgb2hsv(rgb: [number, number, number]): [number, number, number] {
-  const [r, g, b] = rgb;
-  const v = Math.max(r, g, b);
-  const n = v - Math.min(r, g, b);
-
-  // eslint-disable-next-line no-nested-ternary
-  const h = n !== 0 && (v === r ? (g - b) / n : v === g ? 2 + (b - r) / n : 4 + (r - g) / n);
-  // $FlowIgnore
-  return [60 * (h < 0 ? h + 6 : h), v && n / v, v];
-}
-
-// This function mirrors convertCellIdToRGB in the fragment shader of the rendering plane.
-// Output is in [0,1] for H, S, L and A
-export const convertCellIdToHSLA = (id: number, customColors: ?Array<number>): Array<number> => {
-  if (id === 0) {
-    // Return white
-    return [1, 1, 1, 1];
-  }
-
-  let hue;
-
-  if (customColors != null) {
-    const last8Bits = id % 2 ** 8;
-    hue = customColors[last8Bits] || 0;
-  } else {
-    const significantSegmentIndex = id % 2 ** 16;
-
-    const colorCount = 19;
-    const colorIndex = getElementOfPermutation(significantSegmentIndex, colorCount, 2);
-    const colorValueDecimal = (1.0 / colorCount) * colorIndex;
-
-    hue = (1 / 360) * rgb2hsv(colormapJet(colorValueDecimal))[0];
-  }
-
-  return [hue, 1, 0.5, 0.15];
-};
-
 const convertHSLAToCSSString = ([h, s, l, a]) => `hsla(${360 * h}, ${100 * s}%, ${100 * l}%, ${a})`;
 const convertCellIdToCSS = (id, customColors) =>
-  convertHSLAToCSSString(convertCellIdToHSLA(id, customColors));
+  convertHSLAToCSSString(jsConvertCellIdToHSLA(id, customColors));
 
 const hasSegmentation = () => Model.getSegmentationLayer() != null;
